changed type of storeyNo

+++ lib/models/room.dart
@@ -24,7 +24,7 @@ class Room extends BaseModel {
   List<SeepageAnalysis> seepageAnalysis;
   List<StaircaseInspection> staircaseInspection;
   String roomId;
-  int storeyNo;
+  double storeyNo;
   int roomNo;
   List<String> pictures;
   Room({
